+++ src/components/index.js
@@ -1,5 +1,6 @@
 import Vue from 'vue';
 
+import StAutocomplete from './autocomplete/index.vue';
 import StBadge from './badge/index.vue';
 import StButtonGroup from './button-group/index.vue';
 import StButton from './button/index.vue';
@@ -9,6 +10,8 @@ import StCol from './col/index.vue';
 import StCollapser from './collapser/index.vue';
 import StDatepicker from './datepicker/index.vue';
 import StDialog from './dialog/index.vue';
+import StDropdownOption from './dropdown-option/index.vue';
+import StDropdown from './dropdown/index.vue';
 import StIcon from './icon/index.vue';
 import StInput from './input/index.vue';
 import StNotification from './notification/index.vue';
@@ -26,6 +29,7 @@ import StTextarea from './textarea/index.vue';
 import StTooltip from './tooltip/index.vue';
 
 const Components = {
+  StAutocomplete,
   StBadge,
   StButton,
   StButtonGroup,
@@ -35,6 +39,8 @@ const Components = {
   StCollapser,
   StDatepicker,
   StDialog,
+  StDropdown,
+  StDropdownOption,
   StIcon,
   StInput,
   StNotification,