+++ src/components/tabs/_list-default/script.ts
@@ -4,19 +4,12 @@ import {
   Vue,
 } from 'vue-property-decorator';
 
-import StTab from '../_tab/index.vue';
 import { Tab } from '../types';
 
 @Component({
   name: 'StTabsListDefault',
-  components: {
-    StTab,
-  },
 })
 export default class StTabsListDefault extends Vue {
   @Prop(Array)
   tabs!: Tab[];
-
-  @Prop(String)
-  selectedId!: string;
 }